BSIS-829 BSIS-831: change Double/Integer to primitives
Changed ComponentBatch Integer for bloodTransportBoxCount to int
primitive as it provides better performance and there doesn't seem
to be a reason to use the wrapper. Similarily, changed BloodTransportBox
Double for temperature to double.

+++ test/repository/DonationBatchRepositoryTest.java
@@ -162,6 +162,6 @@ public class DonationBatchRepositoryTest extends DBUnitContextDependentTestSuite
   public void testDonationBatchHasComponentBatch() throws Exception {
     DonationBatch one = donationBatchRepository.findDonationBatchById(1l);
     ComponentBatch componentBatch = one.getComponentBatch();
-    Assert.assertEquals("There are 5 boxes on component batch", 5, componentBatch.getBloodTransportBoxCount().intValue());
+    Assert.assertEquals("There are 5 boxes on component batch", 5, componentBatch.getBloodTransportBoxCount());
   }
 }
